restructure first level of pkg
