+++ deeprankcore/neuralnets/naive_gnn.py
