+++ tests/test_integration.py
@@ -3,12 +3,12 @@ from shutil import rmtree
 import os
 import h5py
 from deeprankcore.preprocess import preprocess
-from deeprankcore.models.query import ProteinProteinInterfaceResidueQuery
+from deeprankcore.query import ProteinProteinInterfaceResidueQuery
 from deeprankcore.feature import amino_acid, atomic_contact, biopython, bsa, pssm, sasa
 from tests.utils import PATH_TEST
 from deeprankcore.DataSet import HDF5DataSet
 from deeprankcore.Trainer import Trainer
-from deeprankcore.ginet import GINet
+from deeprankcore.neuralnets.ginet import GINet
 from deeprankcore.models.metrics import OutputExporter
 from deeprankcore.tools.score import get_all_scores
 from deeprankcore.domain.features import nodefeats as Nfeat
