+++ tests/test_preprocess.py
@@ -6,7 +6,7 @@ import importlib
 from typing import List
 import h5py
 from deeprankcore.preprocess import preprocess
-from deeprankcore.models.query import SingleResidueVariantResidueQuery
+from deeprankcore.query import SingleResidueVariantResidueQuery
 from deeprankcore.models.amino_acid import alanine, phenylalanine
 from tests.utils import PATH_TEST
